Add files via upload
So I changed a few things with Heather's code. I added the login (but Jasmin, please replace my code with yours), data storage file (since she mentioned that at the meeting), a "check availability" for the vehicles, and I added tabs for each section (since she mentioned it was too much on one page). Please update the code as needed!